Code split Navbar component
