+++ components/navbar/EventSearcher/index.js
@@ -1,0 +1,1 @@
+export { EventSearcher } from './EventSearcher'
