+++ components/navbar/NavbarComputer/NavbarComputer.css
@@ -1,0 +1,6 @@
+.divider {
+  margin: 0 !important;
+}
+.menu {
+  background: #3A66E5 !important;
+}
