+++ components/navbar/NavbarComputer/index.js
@@ -1,0 +1,1 @@
+export { NavbarComputer } from './NavbarComputer'
