+++ components/navbar/NavbarTabletMobile/index.js
@@ -1,0 +1,1 @@
+export { NavbarTabletMobile } from './NavbarTabletMobile'
