+++ components/navbar/index.js
@@ -1,120 +1,1 @@
-import React from 'react';
-import { Menu, Dropdown, Image, Responsive, Icon } from 'semantic-ui-react';
-
-import { breakpoints } from '../../lib/breakpoints';
-import Logo from '../../assets/universe.svg';
-import { EventSearcher } from './event-searcher';
-
-export const Navbar = (props) => (
-  <React.Fragment>
-    <NavbarComputer {...props} minWidth={breakpoints.computer} />
-    <NavbarTabletMobile {...props} maxWidth={breakpoints.largestTablet} />
-  </React.Fragment>
-);
-
-// size specific, will move these into their own files later
-const NavbarComputer = ({ user, sticky, onManage, minWidth }) => (
-  <Responsive minWidth={minWidth}>
-    <Menu>
-      <Menu.Menu position="left">
-        <UniverseHome />
-        <EventSearcher />
-      </Menu.Menu>
-      <Menu.Menu position="right">
-        <Help loggedIn={!!user} />
-        <FindEvents />
-        <CreateOrHostEvent isHost={user && user.is_host} />
-        <LoginAndSignUp loggedIn={!!user} />
-        <Account user={user} />
-      </Menu.Menu>
-    </Menu>
-  </Responsive>
-);
-
-const NavbarTabletMobile = ({ user, sticky, onManage, maxWidth }) => (
-  <Responsive maxWidth={maxWidth}>
-    <Menu widths={3}>
-      <Menu.Item>
-        <Icon className="universe-search" size="big" />
-      </Menu.Item>
-      <Menu.Item>
-        <Logo />
-      </Menu.Item>
-      <Menu.Item>
-        <Icon className="universe-basic-information" size="big" />
-      </Menu.Item>
-    </Menu>
-  </Responsive>
-);
-
-// private components
-const UniverseHome = () => <Menu.Item href="/"><Logo /></Menu.Item>;
-
-const CreateOrHostEvent = ({ isHost }) => {
-  if (isHost) return <Menu.Item href='/create'>Create Events</Menu.Item>
-  return (
-    <Dropdown item pointing text='Host Events'>
-      <Dropdown.Menu>
-        <Dropdown.Item href="/create">Create Event</Dropdown.Item>
-        <Dropdown.Divider />
-        <Dropdown.Item href="/features">Features</Dropdown.Item>
-        <Dropdown.Item href="/customers">Customers</Dropdown.Item>
-        <Dropdown.Item href="/pricing">Pricing</Dropdown.Item>
-      </Dropdown.Menu>
-    </Dropdown>
-  );
-};
-
-const LoginAndSignUp = ({ loggedIn }) => {
-  if (loggedIn) return null;
-  return (
-    <React.Fragment>
-      <Menu.Item href={`/sign-in?redirectTo=${window.location.href}`}>Log In</Menu.Item>
-      <Menu.Item href={`/users/sign_up?redirectTo=${window.location.href}`}>Sign Up</Menu.Item>
-    </React.Fragment>
-  );
-};
-
-const FindEvents = () => <Menu.Item href="/explore">Find Events</Menu.Item>;
-
-const Help = ({ loggedIn }) => {
-  if (loggedIn) return null;
-  return (
-    <Dropdown item pointing text='Help'>
-      <Dropdown.Menu>
-        <Dropdown.Item href="/help">Help Center</Dropdown.Item>
-        <Dropdown.Divider />
-        <Dropdown.Item href="/help/find_tickets">Where are my tickets?</Dropdown.Item>
-        <Dropdown.Item href="/help/contact_organizer">Contact the event organizer</Dropdown.Item>
-        <Dropdown.Item href="/help/get_refund">Can I get a refund?</Dropdown.Item>
-      </Dropdown.Menu>
-    </Dropdown>
-  );
-};
-
-const Account = ({ user }) => {
-  if (!user) return null;
-
-  const trigger = (
-    <span>
-      <Image avatar src="https://picsum.photos/seed/picsum/50/50" verticalAlign='middle' />
-      {user.name}
-    </span>
-  );
-
-  return (
-    <React.Fragment>
-      <Dropdown item trigger={trigger}>
-        <Dropdown.Menu>
-          <Dropdown.Item href="/dashboard/tickets">Your tickets</Dropdown.Item>
-          <Dropdown.Divider />
-          <Dropdown.Item href={`/users/${user.slug || user.id || ''}`}>Profile</Dropdown.Item>
-          <Dropdown.Item href="/dashboard/conversations">Messages</Dropdown.Item>
-          <Dropdown.Item href="/dashboard/settings">Settings</Dropdown.Item>
-          <Dropdown.Item href="/help">Help</Dropdown.Item>
-          <Dropdown.Item href="/users/sign_out">Log Out</Dropdown.Item>
-        </Dropdown.Menu>
-      </Dropdown>
-    </React.Fragment>
-  );
-};
+export { Navbar } from './Navbar'
